+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Src
+## V0.4.2
 - 🎉 now latex/typst files can also be converted into images / inline images
 - 📈 markdown viewer now handles local images too! (#24) by @Alb-O
 - 📈 markdown viewer now also handles footnotes